nlp-practice/similarity-comparison/requirements.txt
absl-py==0.5.0
astor==0.7.1
boto==2.49.0
boto3==1.9.13
botocore==1.12.13
bz2file==0.98
certifi==2018.8.24
chardet==3.0.4
docutils==0.14
gast==0.2.0
gensim==3.6.0
grpcio==1.15.0
h5py==2.8.0
idna==2.7
jmespath==0.9.3
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
numpy==1.15.2
pandas==0.23.4
protobuf==3.6.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
s3transfer==0.1.13
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
smart-open==1.7.1
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
urllib3==1.23
virtualenv==16.0.0
Werkzeug==0.14.1
matplotlib==3.0.0
Pillow==5.3.0
flask==1.0.2
flask_environments==0.1